14주차 - 최솟값 만들기

+++ BES_3th/src/main/java/week14/MakeMinimum.java
@@ -6,7 +6,7 @@ import java.util.Collections;
 public class MakeMinimum {
 
     public int solution(int[] A, int[] B) {
-        int answer = 0;
+        Integer answer = 0;
         Integer[] tempB = Arrays.stream(B).boxed().toArray(Integer[]::new);
 
         Arrays.sort(A);
@@ -16,7 +16,7 @@ public class MakeMinimum {
             answer += A[i] * tempB[i];
         }
 
-        return (Integer) answer;
+        return answer;
     }
 
 }
